feat: Remove RequirementListings component
This commit removes the RequirementListings component from the frontend. The component was responsible for displaying a list of requirements for a specific project, but it is no longer needed as it has been replaced by the new RequirementViewByProject component. This change improves the organization and presentation of requirement data by using a dedicated page for viewing requirements by project.