refactor: Update fetch request URL in addCompany function
This commit updates the fetch request URL in the addCompany function of the App component. The previous URL was set to "http://localhost:3001/api/v1/companies", but it has been changed to "http://backend:3001/api/v1/companies". This change ensures that the frontend makes API requests to the correct backend URL, improving the reliability and functionality of the application.