refactor: Update fetch request URLs to use relative path
This commit updates the fetch request URLs in the CompanyListings component to use a relative path instead of an absolute URL. By using a relative path, the frontend will make API requests to the correct backend URL, regardless of the environment or container name. This change improves the portability and flexibility of the application.